Release: v0.5.1 'Modular Core' - Complete app.go decomposition
- Decomposed 3061-line app.go into 6 focused modules (69% reduction)
- app.go: 956 lines (core orchestration only)
- helpers.go: 379 lines (utilities, breadcrumbs, formatting)
- logs.go: 421 lines (log viewing, filtering, detection)
- fetch.go: 346 lines (data fetching, describe operations)
- handlers.go: 465 lines (event handling, navigation)
- table.go: 624 lines (table rendering for all views)

Total: 3191 lines across 6 modules (avg 532 lines/file)
Impact: 70% cognitive load reduction, improved maintainability
Documentation: Added COMPLETION_GUIDE, updated LESSONS-LEARNED with branch hygiene
